@@ -33,70 +34,73 @@
   },
   "homepage": "https://github.com/angular/angularfire2#readme",
   "dependencies": {
-    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
-    "@angular-devkit/core": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular-devkit/schematics": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/common": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/compiler": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/core": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-browser": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-browser-dynamic": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/router": ">=6.0.0 <9 || ^9.0.0-0",
-    "firebase": ">= 5.5.7 <8",
-    "firebase-tools": "^6.10.0",
+    "@angular-devkit/architect": "^0.900.0-0 || ^0.900.0",
+    "@angular-devkit/core": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular-devkit/schematics": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/common": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/compiler": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/core": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/platform-browser": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/platform-browser-dynamic": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/router": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "firebase": "^7.6.2",
+    "firebase-tools": "^7.11.0",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
-    "rxjs": "^6.0.0",
-    "ws": "^3.3.2",
+    "rxfire": "^3.9.7",
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
+    "ws": "^7.2.1",
     "xhr2": "^0.1.4",
-    "zone.js": "^0.8.0 || ^0.9.0"
+    "zone.js": "~0.10.2"
   },
   "optionalDependencies": {
-    "bufferutil": "~3.0.0",
-    "utf-8-validate": "~4.0.0"
+    "bufferutil": "^4.0.1",
+    "utf-8-validate": "^5.0.2"
   },
   "devDependencies": {
-    "@angular/animations": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/compiler-cli": ">=6.0.0 <9 || ^9.0.0-0",
-    "@angular/platform-server": ">=6.0.0 <9 || ^9.0.0-0",
+    "@angular-devkit/build-angular": "^0.900.0-0 || ^0.900.0",
+    "@angular-devkit/build-ng-packagr": "^0.900.0-0 || ^0.900.0",
+    "@angular/animations": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/cli": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/compiler-cli": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@angular/platform-server": "^9.0.0-0 || ^9.0.0 || ^10.0.0-0",
+    "@types/fs-extra": "^7.0.0",
+    "@types/gzip-size": "^5.1.1",
     "@types/inquirer": "^0.0.44",
-    "@types/jasmine": "^2.5.36",
+    "@types/jasmine": "^3.3.13",
+    "@types/node": "^12.6.2",
     "@types/request": "0.0.30",
+    "codelyzer": "^5.0.0",
     "concurrently": "^2.2.0",
     "conventional-changelog-cli": "^1.2.0",
-    "es6-module-loader": "^0.17.10",
-    "es6-shim": "^0.35.0",
-    "fs-extra": "^4.0.0",
-    "gulp": "^3.9.0",
-    "gulp-jasmine": "^2.2.1",
-    "gulp-typescript": "^2.10.0",
-    "gzip-size": "^4.0.0",
-    "http-server": "^0.8.5",
-    "jasmine": "^2.4.1",
-    "jasmine-core": "^2.4.1",
-    "json": "^9.0.3",
-    "karma": "^2.0.0",
+    "fs-extra": "^8.0.1",
+    "gzip-size": "^5.1.1",
+    "jasmine": "^3.4.0",
+    "jasmine-core": "^3.4.0",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "^0.3.6",
-    "karma-mocha-reporter": "^2.0.2",
-    "karma-systemjs": "^0.10.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "ncp": "^2.0.0",
-    "parse5": "^1.3.2",
+    "ng-packagr": "^9.0.0-0 || ^9.0.0",
     "pretty-size": "^2.0.0",
     "protractor": "3.0.0",
     "reflect-metadata": "0.1.2",
+    "replace-in-file": "^5.0.2",
     "rimraf": "^2.5.4",
-    "rollup": "^0.64.1",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-watch": "^4.3.1",
-    "schematics-utilities": "^1.1.1",
+    "schematics-utilities": "^2.0.1",
     "shelljs": "^0.8.0",
     "systemjs": "^0.19.16",
     "systemjs-builder": "^0.15.7",
     "traceur": "0.0.96",
-    "typescript": ">=3.4.0 <3.5.0"
+    "tsconfig-paths": "^3.8.0",
+    "tsickle": "^0.35.0",
+    "tslint": "^5.17.0",
+    "typedoc": "^0.16.4",
+    "typescript": "~3.6.4"
   },
   "typings": "index.d.ts"
 }
