@@ -1,31 +1,52 @@
-import { Injectable, InjectionToken } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
+import { Injectable, Inject, Optional, NgZone } from '@angular/core';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { Observable, of, pipe, UnaryFunction } from 'rxjs';
-import { map, switchMap, take } from 'rxjs/operators'
-import { User, auth } from 'firebase/app';
-import { AngularFireAuth } from '@angular/fire/auth';
+import { map, switchMap, take, observeOn, shareReplay } from 'rxjs/operators'
+import { User } from 'firebase/app';
+import { ɵAngularFireSchedulers, FirebaseOptions, FirebaseAppConfig, FIREBASE_OPTIONS, FIREBASE_APP_NAME, ɵfirebaseAppFactory } from '@angular/fire';
 
 export type AuthPipeGenerator = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => AuthPipe;
 export type AuthPipe = UnaryFunction<Observable<User|null>, Observable<boolean|any[]>>;
 
-@Injectable()
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFireAuthGuard implements CanActivate {
 
-  constructor(private afAuth: AngularFireAuth, private router: Router) {}
+  authState: Observable<User|null>;
 
-  canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
-    const authPipeFactory: AuthPipeGenerator = next.data.authGuardPipe || (() => loggedIn);
-    return this.afAuth.user.pipe(
-        take(1),
-        authPipeFactory(next, state),
-        map(canActivate => typeof canActivate == "boolean" ? canActivate : this.router.createUrlTree(canActivate))
+  constructor(
+    @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig:string|FirebaseAppConfig|null|undefined,
+    zone: NgZone,
+    private router: Router
+  ) {
+    const auth = of(undefined).pipe(
+      observeOn(new ɵAngularFireSchedulers(zone).outsideAngular),
+      switchMap(() => zone.runOutsideAngular(() => import('firebase/auth'))),
+      map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
+      map(app => app.auth()),
+      shareReplay({ bufferSize: 1, refCount: false }),
+    );
+
+    this.authState = auth.pipe(
+      switchMap(auth => new Observable<User|null>(auth.onAuthStateChanged.bind(auth)))
+    );
+  }
+
+  canActivate = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => {
+    const authPipeFactory = next.data.authGuardPipe as AuthPipeGenerator || (() => loggedIn);
+    return this.authState.pipe(
+      take(1),
+      authPipeFactory(next, state),
+      map(can => typeof can == "boolean" ? can : this.router.createUrlTree(<any[]>can))
     );
   }
 
 }
 
-export const canActivate = (pipe: AuthPipe|AuthPipeGenerator) => ({
-    canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe.name === "" ? pipe : () => pipe}
+export const canActivate = (pipe: AuthPipeGenerator) => ({
+    canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe }
 });
 
 export const loggedIn: AuthPipe = map(user => !!user);
