@@ -1,16 +1,15 @@
 import { Injectable, Inject, Optional, NgZone, PLATFORM_ID } from '@angular/core';
 import { Observable, of, from } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
-import { FIREBASE_OPTIONS, FIREBASE_APP_NAME, FirebaseOptions, FirebaseAppConfig, FirebaseAuth, _firebaseAppFactory, ɵAngularFireSchedulers, ɵkeepUnstableUntilFirstFactory } from '@angular/fire';
+import { switchMap, map, observeOn, shareReplay } from 'rxjs/operators';
+import { FIREBASE_OPTIONS, FIREBASE_APP_NAME, FirebaseOptions, FirebaseAppConfig, ɵPromiseProxy, ɵlazySDKProxy, ɵfirebaseAppFactory, ɵAngularFireSchedulers, ɵkeepUnstableUntilFirstFactory } from '@angular/fire';
 import { User, auth } from 'firebase/app';
 
-@Injectable()
-export class AngularFireAuth {
+export interface AngularFireAuth extends ɵPromiseProxy<auth.Auth> {};
 
-  /**
-   * Firebase Auth instance
-   */
-  public readonly auth: FirebaseAuth;
+@Injectable({
+  providedIn: 'any'
+})
+export class AngularFireAuth {
 
   /**
    * Observable of authentication state; as of Firebase 4.0 this is only triggered via sign-in/out
@@ -40,28 +39,39 @@ export class AngularFireAuth {
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone
   ) {
-    const keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(new ɵAngularFireSchedulers(zone), platformId);
+    const schedulers = new ɵAngularFireSchedulers(zone);
+    const keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(schedulers, platformId);
 
-    this.auth = zone.runOutsideAngular(() => {
-      const app = _firebaseAppFactory(options, zone, nameOrConfig);
-      return app.auth();
-    });
+    const auth = of(undefined).pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(() => zone.runOutsideAngular(() => import('firebase/auth'))),
+      map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
+      map(app => app.auth()),
+      shareReplay({ bufferSize: 1, refCount: false }),
+    );
 
-    this.authState = new Observable<User | null>(subscriber => {
-      return zone.runOutsideAngular(() => this.auth.onIdTokenChanged(subscriber));
-    }).pipe(keepUnstableUntilFirst);;
+    this.authState = auth.pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(auth => new Observable<User|null>(auth.onAuthStateChanged.bind(auth))),
+      keepUnstableUntilFirst
+    );
 
-    this.user = new Observable<User | null>(subscriber => {
-      return zone.runOutsideAngular(() => this.auth.onIdTokenChanged(subscriber));
-    }).pipe(keepUnstableUntilFirst);
+    this.user = auth.pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(auth => new Observable<User|null>(auth.onIdTokenChanged.bind(auth))),
+      keepUnstableUntilFirst
+    );
 
-    this.idToken = this.user.pipe(switchMap(user => {
-      return user ? from(user.getIdToken()) : of(null)
-    }));
+    this.idToken = this.user.pipe(
+      switchMap(user => user ? from(user.getIdToken()) : of(null))
+    );
 
-    this.idTokenResult = this.user.pipe(switchMap(user => {
-      return user ? from(user.getIdTokenResult()) : of(null)
-    }));
+    this.idTokenResult = this.user.pipe(
+      switchMap(user => user ? from(user.getIdTokenResult()) : of(null))
+    );
+
+    return ɵlazySDKProxy(this, auth, zone);
+
   }
 
 }
