@@ -1,19 +1,24 @@
 import { Injectable, Inject, Optional, NgZone, PLATFORM_ID } from '@angular/core';
-import { isPlatformServer } from '@angular/common';
 import { messaging } from 'firebase/app';
-import { Observable, empty, from, of, throwError } from 'rxjs';
-import { mergeMap, catchError, map, switchMap, concat, defaultIfEmpty, observeOn } from 'rxjs/operators';
-import { FirebaseOptions, FirebaseAppConfig, ɵAngularFireSchedulers, _firebaseAppFactory, FIREBASE_APP_NAME, FIREBASE_OPTIONS } from '@angular/fire';
+import { Observable, empty, of, throwError } from 'rxjs';
+import { mergeMap, catchError, map, switchMap, concat, observeOn, defaultIfEmpty } from 'rxjs/operators';
+import { FirebaseOptions, FirebaseAppConfig, ɵAngularFireSchedulers, FIREBASE_APP_NAME, FIREBASE_OPTIONS, ɵlazySDKProxy, ɵPromiseProxy } from '@angular/fire';
+import { ɵfirebaseAppFactory } from '@angular/fire';
+import { isPlatformServer } from '@angular/common';
 
-@Injectable()
+export interface AngularFireMessaging extends Omit<ɵPromiseProxy<messaging.Messaging>, 'deleteToken'|'getToken'|'requestPermission'> {};
+
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFireMessaging {
-  messaging: Observable<messaging.Messaging>;
-  requestPermission: Observable<void>;
-  getToken: Observable<string|null>;
-  tokenChanges: Observable<string|null>;
-  messages: Observable<{}>;
-  requestToken: Observable<string|null>;
-  deleteToken: (token: string) => Observable<boolean>;
+
+  public readonly requestPermission: Observable<void>;
+  public readonly getToken: Observable<string|null>;
+  public readonly tokenChanges: Observable<string|null>;
+  public readonly messages: Observable<{}>;
+  public readonly requestToken: Observable<string|null>;
+  public readonly deleteToken: (token: string) => Observable<boolean>;
 
   constructor(
     @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
@@ -23,60 +28,63 @@ export class AngularFireMessaging {
   ) {
     const schedulers = new ɵAngularFireSchedulers(zone);
 
-    // @ts-ignore zapping in the UMD in the build script
-    const requireMessaging = from(import('firebase/messaging'));
-
-    this.messaging = requireMessaging.pipe(
+    const messaging = of(undefined).pipe(
       observeOn(schedulers.outsideAngular),
-      map(() => _firebaseAppFactory(options, zone, nameOrConfig)),
-      map(app => app.messaging()),
+      switchMap(() => isPlatformServer(platformId) ? empty() : import('firebase/messaging')),
+      map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
+      map(app => app.messaging())
     );
 
     if (!isPlatformServer(platformId)) {
 
-      this.requestPermission = this.messaging.pipe(
+      this.requestPermission = messaging.pipe(
         observeOn(schedulers.outsideAngular),
         switchMap(messaging => messaging.requestPermission()),
       );
-
+    
     } else {
-
+    
       this.requestPermission = throwError('Not available on server platform.');
-
+    
     }
 
-    this.getToken = this.messaging.pipe(
+    this.getToken = messaging.pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(messaging => messaging.getToken()),
-      defaultIfEmpty(null),
-    );
+      defaultIfEmpty(null)
+    )
 
-    const tokenChanges = this.messaging.pipe(
+    const tokenChanges = messaging.pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(messaging => new Observable(messaging.onTokenRefresh.bind(messaging)).pipe(
         switchMap(() => messaging.getToken())
-      )),
+      ))
     );
 
-    this.tokenChanges = this.getToken.pipe(
+    this.tokenChanges = messaging.pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(messaging => messaging.getToken()),
       concat(tokenChanges)
     );
 
-    this.messages = this.messaging.pipe(
+    this.messages = messaging.pipe(
       observeOn(schedulers.outsideAngular),
-      switchMap(messaging => new Observable(messaging.onMessage.bind(messaging))),
+      switchMap(messaging => new Observable(messaging.onMessage.bind(messaging)))
     );
 
-    this.requestToken = this.requestPermission.pipe(
+    this.requestToken = of(undefined).pipe(
+      switchMap(() => this.requestPermission),
       catchError(() => of(null)),
       mergeMap(() => this.tokenChanges)
     );
 
-    this.deleteToken = (token: string) => this.messaging.pipe(
+    this.deleteToken = (token: string) => messaging.pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(messaging => messaging.deleteToken(token)),
-      defaultIfEmpty(false),
+      defaultIfEmpty(false)
     );
+
+    return ɵlazySDKProxy(this, messaging, zone);
   }
 
 }
