@@ -1,7 +1,7 @@
 import { UploadTaskSnapshot, UploadTask } from './interfaces';
 import { fromTask } from './observable/fromTask';
-import { Observable, from } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 export interface AngularFireUploadTask {
   task: UploadTask,
