@@ -1,9 +1,9 @@
 import { Injectable, Inject, Optional, NgZone, InjectionToken, PLATFORM_ID } from '@angular/core';
-import { of } from 'rxjs';
+import { of, empty, Observable } from 'rxjs';
 import { isPlatformBrowser } from '@angular/common';
 import { map, tap, shareReplay, switchMap, observeOn } from 'rxjs/operators';
-import { FirebaseAppConfig, FirebaseOptions, ɵlazySDKProxy, FirebaseAnalytics, FIREBASE_OPTIONS, FIREBASE_APP_NAME, _firebaseAppFactory, ɵAngularFireSchedulers } from '@angular/fire';
-import { analytics, app } from 'firebase';
+import { FirebaseAppConfig, FirebaseOptions, ɵAngularFireSchedulers, ɵlazySDKProxy, FIREBASE_OPTIONS, FIREBASE_APP_NAME, ɵfirebaseAppFactory, ɵPromiseProxy } from '@angular/fire';
+import { analytics } from 'firebase';
 
 export interface Config {[key:string]: any};
 
@@ -21,28 +21,20 @@ const GTAG_CONFIG_COMMAND = 'config';
 const GTAG_FUNCTION_NAME = 'gtag';
 const DATA_LAYER_NAME = 'dataLayer';
 
-// SEMVER: once we move to Typescript 3.6 use `PromiseProxy<analytics.Analytics>`
-type AnalyticsProxy = {
-  // TODO can we pull the richer types from the Firebase SDK .d.ts? ReturnType<T[K]> is infering
-  //      I could even do this in a manual build-step
-  logEvent(eventName: string, eventParams?: {[key: string]: any}, options?: analytics.AnalyticsCallOptions): Promise<void>,
-  setCurrentScreen(screenName: string, options?: analytics.AnalyticsCallOptions): Promise<void>,
-  setUserId(id: string, options?: analytics.AnalyticsCallOptions): Promise<void>,
-  setUserProperties(properties: analytics.CustomParams, options?: analytics.AnalyticsCallOptions): Promise<void>,
-  setAnalyticsCollectionEnabled(enabled: boolean): Promise<void>,
-  app: Promise<app.App>
-};
+export interface AngularFireAnalytics extends ɵPromiseProxy<analytics.Analytics> {};
 
-export interface AngularFireAnalytics extends AnalyticsProxy {};
+const ANALYTICS_INSTANCE_CACHE = Symbol();
+const ANALYTICS_INITIALIZED = Symbol();
 
-@Injectable()
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFireAnalytics {
 
   private gtag: (...args: any[]) => void;
-  private analyticsInitialized: Promise<void>;
 
   async updateConfig(config: Config) {
-    await this.analyticsInitialized;
+    await global[ANALYTICS_INITIALIZED];
     this.gtag(GTAG_CONFIG_COMMAND, this.options[ANALYTICS_ID_FIELD], { ...config, update: true });
   };
 
@@ -60,23 +52,52 @@ export class AngularFireAnalytics {
 
     const schedulers = new ɵAngularFireSchedulers(zone);
 
+    // Analytics errors if it's not unique from a measurementId standpoint, so we need to cache the instances
+    if (!global[ANALYTICS_INSTANCE_CACHE]) {
+      global[ANALYTICS_INSTANCE_CACHE] = {}
+    };
+
+    let analyticsInitialized: Promise<void> = global[ANALYTICS_INITIALIZED];
+    let analyticsInstanceCache: {[key:string]: Observable<analytics.Analytics>} = global[ANALYTICS_INSTANCE_CACHE];
+    let analytics = analyticsInstanceCache[options[ANALYTICS_ID_FIELD]];
+
     if (isPlatformBrowser(platformId)) {
 
       window[DATA_LAYER_NAME] = window[DATA_LAYER_NAME] || [];
       this.gtag = window[GTAG_FUNCTION_NAME] || function() { window[DATA_LAYER_NAME].push(arguments) }
-      this.analyticsInitialized = zone.runOutsideAngular(() =>
-        new Promise(resolve => {
-          window[GTAG_FUNCTION_NAME] = (...args: any[]) => {
-            if (args[0] == 'js') { resolve() }
-            this.gtag(...args);
-          }
-        })
-      );
+
+      if (!analyticsInitialized) {
+        analyticsInitialized = zone.runOutsideAngular(() =>
+          new Promise(resolve => {
+            window[GTAG_FUNCTION_NAME] = (...args: any[]) => {
+              if (args[0] == 'js') { resolve() }
+              this.gtag(...args);
+            }
+          })
+        );
+      }
 
     } else {
 
-      this.analyticsInitialized = Promise.resolve();
+      analyticsInitialized = Promise.resolve();
       this.gtag = () => {}
+
+    }
+
+    if (!analytics) {
+
+      analytics = of(undefined).pipe(
+        observeOn(schedulers.outsideAngular),
+        switchMap(() => isPlatformBrowser(platformId) ? import('firebase/analytics') : empty()),
+        map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
+        map(app => app.analytics()),
+        tap(analytics => {
+          if (analyticsCollectionEnabled === false) { analytics.setAnalyticsCollectionEnabled(false) }
+        }),
+        shareReplay({ bufferSize: 1, refCount: false }),
+      );
+
+      analyticsInstanceCache[options[ANALYTICS_ID_FIELD]] = analytics;
 
     }
 
@@ -85,21 +106,8 @@ export class AngularFireAnalytics {
     if (providedAppVersion) { this.updateConfig({ [APP_VERSION_KEY]: providedAppVersion }) }
     if (debugModeEnabled)   { this.updateConfig({ [DEBUG_MODE_KEY]:  1 }) }
 
-    const analytics = of(undefined).pipe(
-      observeOn(schedulers.outsideAngular),
-      // @ts-ignore zapping in the UMD in the build script
-      switchMap(() => import('firebase/analytics')),
-      map(() => _firebaseAppFactory(options, zone, nameOrConfig)),
-      // SEMVER no need to cast once we drop older Firebase
-      map(app => <analytics.Analytics>app.analytics()),
-      tap(analytics => {
-        if (analyticsCollectionEnabled === false) { analytics.setAnalyticsCollectionEnabled(false) }
-      }),
-      shareReplay({ bufferSize: 1, refCount: false }),
-    );
-
     return ɵlazySDKProxy(this, analytics, zone);
-    
+
   }
 
 }
