@@ -1,10 +1,10 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { FirebaseApp, AngularFireModule } from '@angular/fire';
-import { COMMON_CONFIG } from './test-config';
-import { AngularFireAuthModule } from '@angular/fire/auth';
-import { AngularFireAuthGuardModule, AngularFireAuthGuard } from '@angular/fire/auth-guard';
+import { COMMON_CONFIG } from '../test-config';
+import { AngularFireAuthGuardModule, AngularFireAuthGuard } from './public_api';
 import { RouterModule, Router } from '@angular/router';
 import { APP_BASE_HREF } from '@angular/common';
+import { rando } from '../firestore/utils.spec';
 
 describe('AngularFireAuthGuard', () => {
     let app: FirebaseApp;
@@ -13,8 +13,7 @@ describe('AngularFireAuthGuard', () => {
     beforeEach(() => {
         TestBed.configureTestingModule({
             imports: [
-                AngularFireModule.initializeApp(COMMON_CONFIG),
-                AngularFireAuthModule,
+                AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
                 AngularFireAuthGuardModule,
                 RouterModule.forRoot([
                     { path: 'a', redirectTo: '/', canActivate: [AngularFireAuthGuard] }
@@ -31,7 +30,7 @@ describe('AngularFireAuthGuard', () => {
     });
 
     afterEach(done => {
-        app.delete().then(done, done.fail);
+        app.delete().then(done, done);
     });
     
     it('should be injectable', () => {
