@@ -1,53 +1,34 @@
-import { InjectionToken, NgModule, Optional, NgZone, VERSION as NG_VERSION, Version } from '@angular/core';
-import { auth, database, messaging, storage, firestore, functions } from 'firebase/app';
-// @ts-ignore (https://github.com/firebase/firebase-js-sdk/pull/1206)
-import firebase from 'firebase/app'; // once fixed can pull in as "default as firebase" above
+import { InjectionToken, NgModule, Optional, NgZone, VERSION as NG_VERSION, Version, PLATFORM_ID, Inject } from '@angular/core';
+import { app, auth, database, messaging, storage, firestore, functions, analytics, performance, remoteConfig } from 'firebase/app';
+import * as firebase from 'firebase/app';
 
-// Public types don't expose FirebaseOptions or FirebaseAppConfig
-export type FirebaseOptions = {[key:string]: any};
-export type FirebaseAppConfig = {[key:string]: any};
+// INVESTIGATE Public types don't expose FirebaseOptions or FirebaseAppConfig, is this the case anylonger?
+export interface FirebaseOptions {[key:string]: any};
+export interface FirebaseAppConfig {[key:string]: any};
 
-// SEMVER drop FirebaseOptionsToken and FirebaseNameOrConfigToken in favor of FIREBASE_OPTIONS and FIREBASE_APP_NAME in next major
-export const FirebaseOptionsToken = new InjectionToken<FirebaseOptions>('angularfire2.app.options');
-export const FirebaseNameOrConfigToken = new InjectionToken<string|FirebaseAppConfig|undefined>('angularfire2.app.nameOrConfig');
-
-export const FIREBASE_OPTIONS = FirebaseOptionsToken;
-export const FIREBASE_APP_NAME = FirebaseNameOrConfigToken;
-
-export type FirebaseDatabase = database.Database;
-export type FirebaseAuth = auth.Auth;
-// SEMVER analytics.Analytics;
-export type FirebaseAnalytics = any;
-export type FirebaseMessaging = messaging.Messaging;
-// SEMVER performance.Performance
-export type FirebasePerformance = any;
-export type FirebaseStorage = storage.Storage;
-export type FirebaseFirestore = firestore.Firestore;
-export type FirebaseFunctions = functions.Functions;
-// SEMVER remoteConfig.RemoteConfig;
-export type FirebaseRemoteConfig = any;
+export const FIREBASE_OPTIONS = new InjectionToken<FirebaseOptions>('angularfire2.app.options');
+export const FIREBASE_APP_NAME = new InjectionToken<string|FirebaseAppConfig|undefined>('angularfire2.app.nameOrConfig');
 
 // Have to implement as we need to return a class from the provider, we should consider exporting
 // this in the firebase/app types as this is our highest risk of breaks
-export class FirebaseApp {
+export class FirebaseApp implements Partial<app.App> {
     name: string;
     options: {};
-    analytics: () => FirebaseAnalytics;
-    auth: () => FirebaseAuth;
-    database: (databaseURL?: string) => FirebaseDatabase;
-    messaging: () => FirebaseMessaging;
-    performance: () => FirebasePerformance;
-    storage: (storageBucket?: string) => FirebaseStorage;
+    analytics: () => analytics.Analytics;
+    auth: () => auth.Auth;
+    database: (databaseURL?: string) => database.Database;
+    messaging: () => messaging.Messaging;
+    performance: () => performance.Performance;
+    storage: (storageBucket?: string) => storage.Storage;
     delete: () => Promise<void>;
-    firestore: () => FirebaseFirestore;
-    functions: (region?: string) => FirebaseFunctions;
-    remoteConfig: () => FirebaseRemoteConfig;
-    registerVersion?: (library: string, version: string) => void;
+    firestore: () => firestore.Firestore;
+    functions: (region?: string) => functions.Functions;
+    remoteConfig: () => remoteConfig.RemoteConfig;
 }
 
 export const VERSION = new Version('ANGULARFIRE2_VERSION');
 
-export function _firebaseAppFactory(options: FirebaseOptions, zone: NgZone, nameOrConfig?: string|FirebaseAppConfig|null) {
+export function ɵfirebaseAppFactory(options: FirebaseOptions, zone: NgZone, nameOrConfig?: string|FirebaseAppConfig|null) {
     const name = typeof nameOrConfig === 'string' && nameOrConfig || '[DEFAULT]';
     const config = typeof nameOrConfig === 'object' && nameOrConfig || {};
     config.name = config.name || name;
@@ -55,17 +36,12 @@ export function _firebaseAppFactory(options: FirebaseOptions, zone: NgZone, name
     const existingApp = firebase.apps.filter(app => app && app.name === config.name)[0] as any;
     // We support FirebaseConfig, initializeApp's public type only accepts string; need to cast as any
     // Could be solved with https://github.com/firebase/firebase-js-sdk/pull/1206
-    const app = (existingApp || zone.runOutsideAngular(() => firebase.initializeApp(options, config as any))) as FirebaseApp;
-    if (app.registerVersion) {
-        app.registerVersion('angularfire', VERSION.full);
-        app.registerVersion('angular', NG_VERSION.full);
-    }
-    return app;
+    return (existingApp || zone.runOutsideAngular(() => firebase.initializeApp(options, config as any))) as FirebaseApp;
 }
 
 const FirebaseAppProvider = {
     provide: FirebaseApp,
-    useFactory: _firebaseAppFactory,
+    useFactory: ɵfirebaseAppFactory,
     deps: [
         FIREBASE_OPTIONS,
         NgZone,
@@ -85,5 +61,9 @@ export class AngularFireModule {
                 { provide: FIREBASE_APP_NAME, useValue: nameOrConfig }
             ]
         }
+    }
+    constructor(@Inject(PLATFORM_ID) platformId:Object ) {
+        firebase.registerVersion('angularfire', VERSION.full, platformId.toString());
+        firebase.registerVersion('angular', NG_VERSION.full);
     }
 }
