@@ -1,8 +1,6 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { PersistenceSettings } from './interfaces';
-import { AngularFirestore, EnablePersistenceToken, PersistenceSettingsToken } from './firestore';
-
-import 'firebase/firestore';
+import { AngularFirestore, ENABLE_PERSISTENCE, PERSISTENCE_SETTINGS } from './firestore';
 
 @NgModule({
   providers: [ AngularFirestore ]
@@ -15,8 +13,8 @@ export class AngularFirestoreModule {
     return {
       ngModule: AngularFirestoreModule,
       providers: [
-        { provide: EnablePersistenceToken, useValue: true },
-        { provide: PersistenceSettingsToken, useValue: persistenceSettings },
+        { provide: ENABLE_PERSISTENCE, useValue: true },
+        { provide: PERSISTENCE_SETTINGS, useValue: persistenceSettings },
       ]
     }
   }
