@@ -1,39 +1,19 @@
 import { InjectionToken, NgZone, PLATFORM_ID, Injectable, Inject, Optional } from '@angular/core';
-
 import { Observable, of, from } from 'rxjs';
-
 import { Settings, PersistenceSettings, CollectionReference, DocumentReference, QueryFn, Query, QueryGroupFn, AssociatedReference } from './interfaces';
 import { AngularFirestoreDocument } from './document/document';
 import { AngularFirestoreCollection } from './collection/collection';
 import { AngularFirestoreCollectionGroup } from './collection-group/collection-group';
-
-import { FirebaseFirestore, FirebaseOptions, FirebaseAppConfig, FIREBASE_OPTIONS, FIREBASE_APP_NAME, _firebaseAppFactory, ɵAngularFireSchedulers, ɵkeepUnstableUntilFirstFactory } from '@angular/fire';
+import { FirebaseOptions, FirebaseAppConfig, FIREBASE_OPTIONS, FIREBASE_APP_NAME, ɵfirebaseAppFactory, ɵAngularFireSchedulers, ɵkeepUnstableUntilFirstFactory } from '@angular/fire';
 import { isPlatformServer } from '@angular/common';
-
-// Workaround for Nodejs build
-// @ts-ignore
-import firebase from 'firebase/app';
-
-// SEMVER: have to import here while we target ng 6, as the version of typescript doesn't allow dynamic import of types
 import { firestore } from 'firebase/app';
 
-// SEMVER drop EnablePersistenceToken, PersistenceSettingsToken, and FirestoreSettingsToken in favor of new export names
 /**
  * The value of this token determines whether or not the firestore will have persistance enabled
  */
-export const EnablePersistenceToken = new InjectionToken<boolean>('angularfire2.enableFirestorePersistence');
-export const PersistenceSettingsToken = new InjectionToken<PersistenceSettings|undefined>('angularfire2.firestore.persistenceSettings');
-export const FirestoreSettingsToken = new InjectionToken<Settings>('angularfire2.firestore.settings');
-
-export const ENABLE_PERSISTENCE = EnablePersistenceToken;
-export const PERSISTENCE_SETTINGS = PersistenceSettingsToken
-export const SETTINGS = FirestoreSettingsToken;
-
-// SEMVER kill this in the next major
-// timestampsInSnapshots was depreciated in 5.8.0
-const major = parseInt(firebase.SDK_VERSION.split('.')[0]);
-const minor = parseInt(firebase.SDK_VERSION.split('.')[1]);
-export const DefaultFirestoreSettings = ((major < 5 || (major == 5 && minor < 8)) ? {timestampsInSnapshots: true} : {}) as Settings;
+export const ENABLE_PERSISTENCE = new InjectionToken<boolean>('angularfire2.enableFirestorePersistence');
+export const PERSISTENCE_SETTINGS = new InjectionToken<PersistenceSettings|undefined>('angularfire2.firestore.persistenceSettings');
+export const SETTINGS = new InjectionToken<Settings>('angularfire2.firestore.settings');
 
 /**
  * A utility methods for associating a collection reference with
@@ -108,9 +88,11 @@ export function associateQuery(collectionRef: CollectionReference, queryFn = ref
  *   }
  * }
  */
-@Injectable()
+@Injectable({
+  providedIn: 'any'
+})
 export class AngularFirestore {
-  public readonly firestore: FirebaseFirestore;
+  public readonly firestore: firestore.Firestore;
   public readonly persistenceEnabled$: Observable<boolean>;
   public readonly schedulers: ɵAngularFireSchedulers;
   public readonly keepUnstableUntilFirst: <T>(obs: Observable<T>) => Observable<T>;
@@ -134,9 +116,10 @@ export class AngularFirestore {
     this.keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(this.schedulers, platformId);
 
     this.firestore = zone.runOutsideAngular(() => {
-      const app = _firebaseAppFactory(options, zone, nameOrConfig);
+      const app = ɵfirebaseAppFactory(options, zone, nameOrConfig);
+      if (!app.firestore) { throw "You must import 'firebase/firestore' before using AngularFirestore" }
       const firestore = app.firestore();
-      firestore.settings(settings || DefaultFirestoreSettings);
+      if (settings) { firestore.settings(settings) }
       return firestore;
     });
 
@@ -184,10 +167,8 @@ export class AngularFirestore {
    * @param queryGroupFn
    */
   collectionGroup<T>(collectionId: string, queryGroupFn?: QueryGroupFn): AngularFirestoreCollectionGroup<T> {
-    if (major < 6) { throw "collection group queries require Firebase JS SDK >= 6.0"}
     const queryFn = queryGroupFn || (ref => ref);
-    const firestore: any = this.firestore; // SEMVER: ditch any once targeting >= 6.0
-    const collectionGroup: Query = firestore.collectionGroup(collectionId);
+    const collectionGroup: Query = this.firestore.collectionGroup(collectionId);
     return new AngularFirestoreCollectionGroup<T>(queryFn(collectionGroup), this);
   }
 
