@@ -1,8 +1,9 @@
 import { ReflectiveInjector, Provider } from '@angular/core';
 import { TestBed, inject } from '@angular/core/testing';
-import { FirebaseApp, FirebaseOptionsToken, AngularFireModule, FirebaseNameOrConfigToken } from '@angular/fire';
-import { AngularFireMessaging, AngularFireMessagingModule } from '@angular/fire/messaging';
-import { COMMON_CONFIG } from './test-config';
+import { FirebaseApp, FIREBASE_OPTIONS, AngularFireModule, FIREBASE_APP_NAME } from '@angular/fire';
+import { AngularFireMessaging, AngularFireMessagingModule } from './public_api';
+import { COMMON_CONFIG } from '../test-config';
+import { rando } from '../firestore/utils.spec';
 
 describe('AngularFireMessaging', () => {
   let app: FirebaseApp;
@@ -11,7 +12,7 @@ describe('AngularFireMessaging', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
         AngularFireMessagingModule
       ]
     });
@@ -21,9 +22,8 @@ describe('AngularFireMessaging', () => {
     })();
   });
 
-  afterEach(done => {
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   it('should be exist', () => {
@@ -31,7 +31,7 @@ describe('AngularFireMessaging', () => {
   });
 
   it('should have the FCM instance', () => {
-    expect(afm.messaging).toBeDefined();
+    expect(afm.deleteToken).toBeDefined();
   });
 
 });
@@ -45,12 +45,12 @@ describe('AngularFireMessaging with different app', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG),
+        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
         AngularFireMessagingModule
       ],
       providers: [
-        { provide: FirebaseNameOrConfigToken, useValue: FIREBASE_APP_NAME_TOO },
-        { provide: FirebaseOptionsToken, useValue: COMMON_CONFIG }
+        { provide: FIREBASE_APP_NAME, useValue: FIREBASE_APP_NAME_TOO },
+        { provide: FIREBASE_OPTIONS, useValue: COMMON_CONFIG }
       ]
     });
     inject([FirebaseApp, AngularFireMessaging], (app_: FirebaseApp, _afm: AngularFireMessaging) => {
@@ -59,9 +59,8 @@ describe('AngularFireMessaging with different app', () => {
     })();
   });
 
-  afterEach(done => {
+  afterEach(() => {
     app.delete();
-    done();
   });
 
   describe('<constructor>', () => {
@@ -71,7 +70,7 @@ describe('AngularFireMessaging with different app', () => {
     });
 
     it('should have the FCM instance', () => {
-      expect(afm.messaging).toBeDefined();
+      expect(afm.deleteToken).toBeDefined();
     });
 
   });
